Refactory and change test classses.
