+++ src/LineBot/ILineBot.cs
@@ -26,9 +26,9 @@ namespace Line
         /// <summary>
         /// Creates a rich menu.
         /// </summary>
-        /// <param name="richMenuRequest">The rich menu represented as a rich menu object.</param>
+        /// <param name="richMenu">The rich menu represented as a rich menu object.</param>
         /// <returns>.</returns>
-        Task<string> CreateRichMenu(IRichMenuRequest richMenuRequest);
+        Task<string> CreateRichMenu(IRichMenu richMenu);
 
         /// <summary>
         /// Returns the content of the specified message.
