+++ src/LineBot/RichMenu/IRichMenu.cs
@@ -12,18 +12,40 @@
 // License for the specific language governing permissions and limitations
 // under the License.
 
-using Newtonsoft.Json;
-
 namespace Line
 {
     /// <summary>
-    /// Rich menu response object with the rich menu ID. This object is returned when you get a rich menu or get a list of rich menus.
+    /// Encapsulates a rich menu request.
     /// </summary>
-    internal interface IRichMenu
+    public interface IRichMenu
     {
         /// <summary>
-        /// Gets the rich menu ID.
+        /// Gets the object which contains the width and height of the rich menu displayed in the chat.
+        /// Rich menu images must be one of the following sizes: 2500x1686px or 2500x843px.
         /// </summary>
-        string RichMenuId { get; }
+        IRichMenuSize Size { get; }
+
+        /// <summary>
+        /// Gets a value indicating whether the rich menu should be displayed by default.
+        /// </summary>
+        bool Selected { get; }
+
+        /// <summary>
+        /// Gets the name of the rich menu. This value can be used to help manage your rich menus and is not displayed to users.
+        /// Max: 300 characters.
+        /// </summary>
+        string Name { get; }
+
+        /// <summary>
+        /// Gets the text displayed in the chat bar.
+        /// Max: 14 characters.
+        /// </summary>
+        string ChatBarText { get; }
+
+        /// <summary>
+        /// Gets the array of area objects which define the coordinates and size of tappable areas
+        /// Max: 20 area objects.
+        /// </summary>
+        IRichMenuArea[] Areas { get; }
     }
 }
