--- tests/LineBot.Tests/RichMenuRequestTests/TheRichMenuBoundsClassTests.cs
@@ -1,201 +1,0 @@
-﻿// Copyright 2017-2018 Dirk Lemstra (https://github.com/dlemstra/line-bot-sdk-dotnet)
-//
-// Dirk Lemstra licenses this file to you under the Apache License,
-// version 2.0 (the "License"); you may not use this file except in compliance
-// with the License. You may obtain a copy of the License at:
-//
-//   https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations
-// under the License.
-
-using System;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Line.Tests
-{
-    [TestClass]
-    public class TheRichMenuBoundsClassTests
-    {
-        [TestMethod]
-        public void ShouldThrowExceptionWhenXIsBiggerThan2500()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The horizontal position cannot be bigger than 2500.", () =>
-            {
-                richMenuBounds.X = 2501;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenXIs2500()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                X = 2500
-            };
-
-            Assert.AreEqual(2500, richMenuBounds.X);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenXIsLessThan0()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The horizontal position cannot be less than 0.", () =>
-            {
-                richMenuBounds.X = -1;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenXIs0()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                X = 0
-            };
-
-            Assert.AreEqual(0, richMenuBounds.X);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenYIsBiggerThan1686()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The vertical position cannot be bigger than 1686.", () =>
-            {
-                richMenuBounds.Y = 1687;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenYIs1686()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Y = 1686
-            };
-
-            Assert.AreEqual(1686, richMenuBounds.Y);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenYIsLessThan0()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The vertical position cannot be less than 0.", () =>
-            {
-                richMenuBounds.Y = -1;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenYIs0()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Y = 0
-            };
-
-            Assert.AreEqual(0, richMenuBounds.Y);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenWidthIsBiggerThan2500()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The width cannot be bigger than 2500.", () =>
-            {
-                richMenuBounds.Width = 2501;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenWidthIs2500()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Width = 2500
-            };
-
-            Assert.AreEqual(2500, richMenuBounds.Width);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenWidthIsLessThan1()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The width cannot be less than 1.", () =>
-            {
-                richMenuBounds.Width = 0;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenWidthIs1()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Width = 1
-            };
-
-            Assert.AreEqual(1, richMenuBounds.Width);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenHeightIsBiggerThan1686()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The height cannot be bigger than 1686.", () =>
-            {
-                richMenuBounds.Height = 1687;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenheightIs1686()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Height = 1686
-            };
-
-            Assert.AreEqual(1686, richMenuBounds.Height);
-        }
-
-        [TestMethod]
-        public void ShouldThrowExceptionWhenHeightIsLessThan1()
-        {
-            var richMenuBounds = new RichMenuBounds();
-
-            ExceptionAssert.Throws<InvalidOperationException>("The height cannot be less than 1.", () =>
-            {
-                richMenuBounds.Height = 0;
-            });
-        }
-
-        [TestMethod]
-        public void ShouldNotThrowExceptionWhenheightIs1()
-        {
-            var richMenuBounds = new RichMenuBounds()
-            {
-                Height = 1
-            };
-
-            Assert.AreEqual(1, richMenuBounds.Height);
-        }
-    }
-}
